Add oem structure
- Add oem as a redfish submodule.
- Modify code and files to avoid circular imports :
	- types.py contains the Base, BaseCollection and Device class on
	  which all standard and oem classes are inherited.
	- standard.py will contain the redfish standard classes.
	- oem/<oem_name> will contains specific classes for <oem_name>.
          so oem/hpe will contains hpe specific classes.
	  Note: the goal of python-redfish is not to deal with oem part.
	  So oem will be kept as minimal as possible, however currently
	  some critical hardware characteristics are only available into the oem
	  part that's the reason why we are implementing it.
