+++ setup.cfg
@@ -23,6 +23,7 @@ classifier =
 [files]
 packages =
    redfish
+   redfish/oem
 
 scripts =
 	redfish-client/redfish-client
